requirements.txt
# Core Dependencies
Flask==3.0.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9

# SOAP Client
zeep==4.2.1
requests==2.31.0

# Azure Event Hub
azure-eventhub==5.11.6
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging and Utilities
python-json-logger==2.0.7

Werkzeug==3.0.1